+++ src/Resources/Resources/ChangeLog.md
@@ -24,7 +24,8 @@
 * Updated `Export-AzResourceGroup` to use the SDK.
 * Added culture info to parse methods
 * Changed Double parser for version parser
-* Changed New-RoleAssignment to include principal type during calls
+* Changed New-AzRoleAssignment to include principal type during calls
+* Changed Set-AzRoleAssignment to include principal type during calls
 
 ## Version 3.0.0
 * Fixed parsing bug